# removed deprecated .NET versions & performance improvements