--- a/TwitchLib.EventSub.Webhooks.Test/UnitTest1.cs
+++ b/TwitchLib.EventSub.Webhooks.Test/UnitTest1.cs
@@ -13,7 +13,7 @@ public class UnitTest1
         var messageBody = """
                     {"subscription":{"id":"d5b930eb-5333-a535-0e38-603bc2a22ce1","status":"enabled","type":"stream.offline","version":"1","condition":{"broadcaster_user_id":"88922321"},"transport":{"method":"webhook","callback":"null"},"created_at":"2025-05-11T20:13:00.4746258Z","cost":0},"event":{"broadcaster_user_id":"88922321","broadcaster_user_login":"testBroadcaster","broadcaster_user_name":"testBroadcaster"}}
                     """;
-        var secret = "supersecuresecret"u8.ToArray();
+        var secret = "supersecuresecret"u8;
 
         Assert.True(EventSubSignatureVerificationMiddleware.IsSignatureValid(messageSignature, messageId, messageTimestamp, messageBody, secret));
     }