--- a/TwitchLib.EventSub.Webhooks.Test/xunit.runner.json
+++ b/TwitchLib.EventSub.Webhooks.Test/xunit.runner.json
@@ -1,0 +1,3 @@
+{
+    "$schema": "https://xunit.net/schema/current/xunit.runner.schema.json"
+}