--- a/TwitchLib.EventSub.Webhooks.sln
+++ b/TwitchLib.EventSub.Webhooks.sln
@@ -1,11 +1,13 @@
 ﻿
 Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 16
-VisualStudioVersion = 16.0.31410.223
+# Visual Studio Version 17
+VisualStudioVersion = 17.13.35931.197
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "TwitchLib.EventSub.Webhooks", "TwitchLib.EventSub.Webhooks\TwitchLib.EventSub.Webhooks.csproj", "{6176B27E-E417-41C8-BDEE-CB09FD342CA5}"
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "TwitchLib.EventSub.Webhooks.Example", "TwitchLib.EventSub.Webhooks.Example\TwitchLib.EventSub.Webhooks.Example.csproj", "{26079F7D-9C72-43F9-BCAC-AD7E96AAC5C3}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "TwitchLib.EventSub.Webhooks.Test", "TwitchLib.EventSub.Webhooks.Test\TwitchLib.EventSub.Webhooks.Test.csproj", "{6951CCDA-242A-8652-D81D-F20D8EF8FD9F}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -21,6 +23,10 @@ Global
 		{26079F7D-9C72-43F9-BCAC-AD7E96AAC5C3}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{26079F7D-9C72-43F9-BCAC-AD7E96AAC5C3}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{26079F7D-9C72-43F9-BCAC-AD7E96AAC5C3}.Release|Any CPU.Build.0 = Release|Any CPU
+		{6951CCDA-242A-8652-D81D-F20D8EF8FD9F}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{6951CCDA-242A-8652-D81D-F20D8EF8FD9F}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{6951CCDA-242A-8652-D81D-F20D8EF8FD9F}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{6951CCDA-242A-8652-D81D-F20D8EF8FD9F}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE