--- a/TwitchLib.EventSub.Webhooks/AssemblyInfo.cs
+++ b/TwitchLib.EventSub.Webhooks/AssemblyInfo.cs
@@ -1,0 +1,3 @@
+﻿using System.Runtime.CompilerServices;
+
+[assembly:InternalsVisibleTo("TwitchLib.EventSub.Webhooks.Test")]