--- a/TwitchLib.EventSub.Webhooks/Middlewares/EventSubNotificationLoggerMiddleware.cs
+++ b/TwitchLib.EventSub.Webhooks/Middlewares/EventSubNotificationLoggerMiddleware.cs
@@ -20,11 +20,10 @@ namespace TwitchLib.EventSub.Webhooks.Middlewares
 
         public async Task InvokeAsync(HttpContext context)
         {
-            var stopwatch = new Stopwatch();
-            stopwatch.Start();
+            var startTimestamp= Stopwatch.GetTimestamp();
             await _next(context);
-            stopwatch.Stop();
-            _logger.LogEventSubHttpNotification(context.Request.Path, context.Response.StatusCode, stopwatch.Elapsed.TotalMilliseconds);
+            var elapsed = Stopwatch.GetElapsedTime(startTimestamp);
+            _logger.LogEventSubHttpNotification(context.Request.Path, context.Response.StatusCode, elapsed.TotalMilliseconds);
         }
     }
 }