--- a/TwitchLib.EventSub.Webhooks/TwitchLib.EventSub.Webhooks.csproj
+++ b/TwitchLib.EventSub.Webhooks/TwitchLib.EventSub.Webhooks.csproj
@@ -1,7 +1,7 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk">
 
   <PropertyGroup>
-    <TargetFrameworks>net5.0;net6.0;net7.0;net8.0</TargetFrameworks>
+    <TargetFrameworks>net8.0</TargetFrameworks>
     <PackageId>TwitchLib.EventSub.Webhooks</PackageId>
     <Authors>swiftyspiffy, Prom3theu5, Syzuna, LuckyNoS7evin</Authors>
     <PackageTags>twitchlib;twitch,eventsub;webhooks</PackageTags>
@@ -30,7 +30,6 @@
 
   <ItemGroup>
     <PackageReference Include="Microsoft.Extensions.Logging" Version="8.0.0" />
-    <PackageReference Include="System.Text.Json" Version="8.0.0" />
     <PackageReference Include="TwitchLib.EventSub.Core" Version="2.5.3-preview-558d8a5" />
   </ItemGroup>