This PR is made from 2 commits.
In the first commit, I only modified the code to be able to test the signature verification function (`IsSignatureValid`).
In the second commit, I made changes to improve performance (below is a benchmark that measures the performance of `IsSignatureValid` before and after the changes) and removed deprecated versions of .NET.

```
| Method                  | Mean       | Error    | StdDev   | Gen0   | Allocated |
|------------------------ |-----------:|---------:|---------:|-------:|----------:|
| IsSignatureValid_Old    | 3,351.4 ns | 63.17 ns | 59.09 ns | 0.0725 |    4424 B |
| IsSignatureValid_New    |   906.9 ns | 16.50 ns | 15.44 ns |      - |      56 B |
```